Modified the background subtraction algorithm to only keep continuous blocks of foreground pixels (such that the continuous block contains some percentage of all foreground pixels).
Also started a class for performing the sequence of preprocessing
steps:
	1. remove background, extract foreground pixels
	2. map to a lower dimension using PCA
	3. cluster the images in a lower dimension using kmeans
These steps can be trained a set of images and then the same
transformations can be applied to a new test dataset.

I plan to change the clustering portion to include more than just kmeans
(likely it needs something that can handle non-linear cluster shape).

Also need to decide if the foreground should be simplified such that
the clustering is only done on a set of images that represent the poses.