Combined the random translation and rotation functions into a single image distortion function. Also added another component to the image distortions that translates the image randomly.
Began running the random distortions functions on the full data set,
discovered some bugs that cropped the images, but everything
seems to be working now.

Still need to write the final wrapper function around the random
image distortions to build out a large dataset and save the target
variable labels for each image along the way.

Currently the code is a little slow. There's a lot that goes on
to distort the images, luckily this will only have to be run once.